+++ spec/f_service/result/failure_spec.rb
@@ -9,8 +9,15 @@ RSpec.describe FService::Result::Failure do
   it { expect(failure).to be_failed }
   it { expect(failure).not_to be_successful }
   it { expect(failure.error).to eq('Whoops!') }
+  it { expect(failure.type).to eq(nil) }
   it { expect(failure.value).to eq(nil) }
   it { expect { failure.value! }.to raise_error FService::Result::Error }
+
+  context 'when defining a type' do
+    subject(:failure) { described_class.new('Whoops!', :error) }
+
+    it { expect(failure.type).to eq(:error) }
+  end
 
   describe '#on' do
     context 'when matching results' do