+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased (master)
 ### Added
+- Allow adding types on `Result`s.
 - Add `#on_success` and `#on_failure` hooks on Result objects.
 - Link to Changelog on gemspec.